+++ server/api/orders.js
@@ -1,0 +1,23 @@
+const router = require('express').Router()
+const {Order} = require('../db/models')
+module.exports = router
+
+// GET /api/orders
+router.get('/', async (req, res, next) => {
+  try {
+    const orders = await Order.findAll()
+    res.json(orders)
+  } catch (error) {
+    next(error)
+  }
+})
+
+// GET /api/orders/:orderId
+router.get('/:orderId', async (req, res, next) => {
+  try {
+    const order = await Order.findByPk(req.params.orderId)
+    res.json(order)
+  } catch (error) {
+    next(error)
+  }
+})
